New Composite PolygonLayer. Address GeoJson/Polygon Layer audit
