+++ src/layers/core/solid-polygon-layer/polygon-tesselator-extruded.js
