+++ src/layers/core/solid-polygon-layer/polygon-tesselator.js
