+++ src/layers/core/solid-polygon-layer/polygon.js
