+++ src/layers/core/solid-polygon-layer/solid-polygon-layer-fragment.glsl
