+++ src/layers/core/solid-polygon-layer/solid-polygon-layer-vertex-64.glsl
