+++ src/layers/core/solid-polygon-layer/solid-polygon-layer-vertex.glsl
